pyproject.toml
[project]
name = "opaquetoolsbench"
version = "0.1.0"
description = "Tool use evaluation and analysis project"
requires-python = "==3.10"
dependencies = [
    # Additional packages for your analysis scripts
    "pandas>=2.0.0",
    "numpy>=1.20.0",
    "matplotlib>=3.5.0",
    # From requirements.txt
    "python-chess>=1.0.0",
    "requests>=2.31.0",
    "torch>=2.0.0",
    "sentencepiece>=0.1.99",
    "protobuf>=4.25.1",
    # Original BrowseComp-Plus dependencies
    "accelerate>=1.9.0",
    "anthropic>=0.58.2",
    "datasets>=4.0.0",
    "deepspeed>=0.17.2",
    "faiss-cpu>=1.11.0.post1",
    "fastmcp==2.9.2",
    "google-genai>=1.27.0",
    "openai>=1.97.0",
    "peft>=0.16.0",
    "pyngrok>=7.2.12",
    "pyserini>=1.2.0",
    "python-dotenv>=1.1.1",
    "qwen-agent[gui,mcp,rag]==0.0.27",
    "qwen-omni-utils==0.0.8",
    "rich>=14.0.0",
    "tevatron",
    "torchvision",
    "tqdm>=4.67.1",
    "transformers>=4.53.2",
    "vllm>=0.9.0",
    "tree-sitter==0.21.3",
    "tree-sitter-java==0.21.0",
    "tree-sitter-javascript==0.21.4",
    "cohere==5.5.8",
    "mistralai==0.4.2",
    "together>=1.5.25",
]

[tool.uv.sources]
tevatron = { git = "https://github.com/texttron/tevatron.git", rev = "main" }
